new db structure and new import

+++ backend/src/main/java/ru/backend/model/Combo.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -23,9 +24,11 @@ public class Combo {
     private String title;
 
     @Setter
-    private Long price;
+    private BigDecimal price;
 
 //    @OneToMany(mappedBy = "combo", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+//    private List<ComboSlot> slots = new ArrayList<>();
+//    @OneToMany(mappedBy = "combo",  cascade = CascadeType.ALL)
 //    private List<ComboSlot> slots = new ArrayList<>();
     @OneToMany(mappedBy = "combo", cascade = CascadeType.ALL, orphanRemoval = true)
     private Set<ComboSlot> slots = new HashSet<>();
